+++ sr_ronex_drivers/cfg/GeneralIO.cfg
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-PACKAGE="sr_ronex_ethercat_drivers"
+PACKAGE="sr_ronex_drivers"
 
 from dynamic_reconfigure.parameter_generator_catkin import *
