+++ sr_ronex_drivers/ethercat_device_plugin.xml
@@ -1,29 +1,29 @@
-<library path="lib/libsr_ronex_ethercat_drivers" >
+<library path="lib/libsr_ronex_drivers" >
 
   <!--
       The following definitions are EthercatDevice plugins (they are the drivers for a certain type of ethercat slave device, determined by its product code)
   -->
-  <class name="sr_ronex_ethercat_drivers/87032868" type="SrBoard0X" base_class_type="EthercatDevice">
+  <class name="sr_ronex_drivers/87032868" type="SrBoard0X" base_class_type="EthercatDevice">
     <description>
       SrBoard0X - Shadow Robot RoNeX Board
     </description>
   </class>
-  <class name="sr_ronex_ethercat_drivers/33554433" type="SrBoardMk2GIO" base_class_type="EthercatDevice">
+  <class name="sr_ronex_drivers/33554433" type="SrBoardMk2GIO" base_class_type="EthercatDevice">
     <description>
       RoNeX mk2 - General I/O board. Product ID is 0x02000001.
     </description>
   </class>
-  <class name="sr_ronex_ethercat_drivers/33554432" type="StandardEthercatDevice" base_class_type="EthercatDevice">
+  <class name="sr_ronex_drivers/33554432" type="StandardEthercatDevice" base_class_type="EthercatDevice">
     <description>
       RoNeX bridge - Product ID is 0x02000000.
     </description>
   </class>
-  <class name="sr_ronex_ethercat_drivers/0" type="StandardEthercatDevice" base_class_type="EthercatDevice">
+  <class name="sr_ronex_drivers/0" type="StandardEthercatDevice" base_class_type="EthercatDevice">
     <description>
       RoNeX bridge - Kept for compatibility reason. To be removed once we've fixed RoNeX bridge issue.
     </description>
   </class>
-  <class name="sr_ronex_ethercat_drivers/7" type="StandardEthercatDevice" base_class_type="EthercatDevice">
+  <class name="sr_ronex_drivers/7" type="StandardEthercatDevice" base_class_type="EthercatDevice">
     <description>
       RoNeX bridge - duplicated for backward compatibility issues.
     </description>
@@ -33,7 +33,7 @@
       The following definitions are CodDecod plugins (they encode decode information from a certain (unique) ethercat slave device, determined by its product code and serial number)
   -->
 
-  <class name="sr_ronex_ethercat_drivers/87032868_0" type="sr_cod_decod::CodDecodStdIo" base_class_type="sr_cod_decod::CodDecod">
+  <class name="sr_ronex_drivers/87032868_0" type="sr_cod_decod::CodDecodStdIo" base_class_type="sr_cod_decod::CodDecod">
     <description>
       CodDecodStdIo will be used as the default cod_decod plugin to be loaded when a specific plugin for a certain serial number is not found
     </description>
