+++ sr_ronex_drivers/include/sr_ronex_drivers/cod_decod/cod_decod_manager.hpp
@@ -22,7 +22,7 @@
 
 #include <boost/smart_ptr.hpp>
 #include <pr2_hardware_interface/hardware_interface.h>
-#include "sr_ronex_ethercat_drivers/cod_decod/cod_decod.hpp"
+#include "sr_ronex_drivers/cod_decod/cod_decod.hpp"
 #include <al/ethercat_slave_handler.h>
 #include <pluginlib/class_loader.h>
