+++ sr_ronex_drivers/test/test_ethercat_drivers.test
@@ -1,3 +1,3 @@
 <launch>
-  <test test-name="test_ethercat_drivers" pkg="sr_ronex_ethercat_drivers" type="test_ethercat_drivers"/>
+  <test test-name="test_ethercat_drivers" pkg="sr_ronex_drivers" type="test_ethercat_drivers"/>
 </launch>
